+++ extensions/terminal-suggest/src/shell/common.ts
@@ -94,49 +94,3 @@ export async function getAliasesHelper(command: string, args: string[], regex: R
 	}
 	return result;
 }
-export async function getZshBashBuiltins(
-	options: ExecOptionsWithStringEncoding,
-	scriptToRun: string,
-	getInfo: string | ((description: string) => string | undefined),
-	existingCommands?: Set<string>,
-): Promise<(string | ICompletionResource)[]> {
-	const compgenOutput = await execHelper(scriptToRun, options);
-	const filter = (cmd: string) => cmd && !existingCommands?.has(cmd);
-	const builtins: string[] = compgenOutput.split('\n').filter(filter);
-	const completions: ICompletionResource[] = [];
-	if (builtins.find(r => r === '.')) {
-		completions.push({
-			label: '.',
-			detail: 'Source a file in the current shell',
-			kind: vscode.TerminalCompletionItemKind.Method
-		});
-	}
-
-	for (const cmd of builtins) {
-		if (typeof cmd === 'string') {
-			try {
-				let description;
-				if (typeof getInfo === 'string') {
-					description = (await execHelper(`${getInfo} ${cmd}`, options))?.trim();
-				} else {
-					description = getInfo(cmd);
-				}
-				completions.push({
-					label: cmd,
-					documentation: description,
-					kind: vscode.TerminalCompletionItemKind.Method
-				});
-
-			} catch (e) {
-				// Ignore errors
-				console.log(`Error getting info for ${e}`);
-				completions.push({
-					label: cmd,
-					kind: vscode.TerminalCompletionItemKind.Method
-				});
-			}
-		}
-	}
-
-	return completions;
-}
